add separate component for DotPlot
